Update README.md
A Revenue Generating System is a structured process or framework that an organization, government, or business uses to collect, manage, and allocate funds. In the context of governments, a revenue system refers to the methods and channels through which a government generates income to fund its operations and public services.